Add files via upload
I have spent about 4 hours on this project. I am having trouble making the destroyer script work and the road just keeps going making the game bigger and bigger. I am also curious if there is way to only make the player score if the car is on the yellow line.